Fix end-session leaving an interactive game running; default to short; denoise music
1. Ending a session (farewell page) previously only cleared the session
   state and left the same card/draw button fully interactive, silently
   falling back to classic endless-draw mode — tapping the screen kept
   drawing new cards past the chosen length, which is what got reported.
   Now it reloads the page: back to the welcome screen with everything
   (deck, stats, session) genuinely reset, rather than hand-resetting
   every piece of state individually.

2. Default game length is now "قصيرة" instead of "طويلة" — a quick
   session is the more sensible first-run default; changed in both the
   game state and the length menu's initial active item.

3. Denoised bg-music.mp3 (ffmpeg afftdn) — the source recording's own
   background hiss becomes noticeable once looped continuously in a
   quiet room. Re-derived from the original source (not the already-
   trimmed file) with the same fade-in/out reapplied after denoising.
   Verified: spectral flatness at the track's quietest passage dropped
   from 0.047 to 0.023 (roughly half the noise-to-tonal ratio) while
   the actual music level at that point was unchanged (-38.2dB vs
   -38.4dB), confirming the hiss was reduced without dulling the music.

Verified 1 and 2 end-to-end: menu defaults to قصيرة (70 cards), deck
exhausts and shows the farewell page at exactly 70 draws, "end session"
triggers a real navigation and lands back on the welcome page.

+++ src/game.ts
@@ -185,8 +185,8 @@ function setCardText(cardText: HTMLElement, text: string): void {
 }
 
 export function initGame(el: Elements, timer: Timer, onEndSession: () => void): Game {
-  // Defaults to the full bank; set from the "⋮" length menu.
-  let gameLength: GameLength = 'long';
+  // Defaults to a quick session; set from the "⋮" length menu.
+  let gameLength: GameLength = 'short';
 
   // Independent of length: a pure single-category mode for a calmer,
   // no-points, no-challenges session — bypasses the length ratios entirely
@@ -423,7 +423,7 @@ export function initGame(el: Elements, timer: Timer, onEndSession: () => void): 
     // A dot on the "⋮" button is the only always-visible cue once a
     // non-default choice is active, since the menu itself stays closed
     // the rest of the time.
-    el.lengthToggle.classList.toggle('has-filter', calmOnly || gameLength !== 'long');
+    el.lengthToggle.classList.toggle('has-filter', calmOnly || gameLength !== 'short');
   }
 
   el.lengthMenu.addEventListener('click', (e) => {
